david. AutoattackPlayer, arrows del player, takeDamages de enemigos y player. refactor SpawnEnemies

+++ Assets/Scripts/Scriptables/EnemyScriptable.cs
@@ -13,7 +13,7 @@ namespace EnemySystem
         public Sprite sprite;
 
         [Header("STATS")]
-        public float hpMax;
+        public int hpMax;
         [Space]
         public float physicalArmor;
         public float magicalArmor;
